[#189] - 입단증 보기 문구 변경
